Create README.md
JobVia aims to simplify the process of finding and booking services while providing a reliable and trustworthy platform for service providers to showcase their skills. Whether you're a customer looking for a handyman or a service provider looking to expand your clientele, ServiceHub is your go-to destination for all your service needs.